requirements.txt
neo4j==5.14.0
sentence-transformers
pyyaml==6.0.1
numpy